Updating API  - added APIException to give you better error handling and removing message hash authentication
No longer checking for message hash signature
Added APIException class which is thrown from all of the methods in case
of error. It will contain a JSON message from the server and, if
available, the full HttpWebResponse object as well. For 409 conflict
responses, check the response header Retry-After for # of seconds to
wait until you can retry.
Updated Readme to show an example of error handling as well.